requirements.txt
openai 
flask 
requests 
streamlit
python-dotenv
flask-cors